-- making link to public --

+++ app/models/fyle.rb
@@ -29,6 +29,7 @@ class Fyle
   end
   
   def is_public?
+    return false unless File.exists?( htaccess_path )
     old_htaccess = ( File.readlines( htaccess_path ) || [] ) * ''
     old_htaccess.match( make_public_string )
   end
@@ -43,6 +44,7 @@ class Fyle
   end
   
   def unmake_public
+    return true unless File.exists?( htaccess_path )
     old_htaccess = ( File.readlines( htaccess_path ) || [] ) * ''
     File.open( htaccess_path, 'w' ) { |f| f.write old_htaccess.gsub( make_public_string, '' ) }  if old_htaccess.match( make_public_string )
   end
